moderatsiya-kartochek-5706/requirements.txt
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sympy==1.13.0
torch==2.2.2
torchvision==0.17.2
tqdm
